Move MainWrapper and use in Staking
