+++ frontend/src/routes/(app)/(u)/(list)/accounts/+layout.svelte
@@ -1,9 +1,9 @@
 <script lang="ts">
   import { goto } from "$app/navigation";
   import Content from "$lib/components/layout/Content.svelte";
+  import IslandWidthMain from "$lib/components/layout/IslandWidthMain.svelte";
   import Layout from "$lib/components/layout/Layout.svelte";
   import LayoutList from "$lib/components/layout/LayoutList.svelte";
-  import MainWrapper from "$lib/components/tokens/MainWrapper.svelte";
   import { AppPath } from "$lib/constants/routes.constants";
   import { accountsTitleStore } from "$lib/derived/accounts-title.derived";
 
@@ -13,9 +13,9 @@
 <LayoutList title={$accountsTitleStore}>
   <Layout>
     <Content {back}>
-      <MainWrapper>
+      <IslandWidthMain>
         <slot />
-      </MainWrapper>
+      </IslandWidthMain>
     </Content>
   </Layout>
 </LayoutList>
